backend/requirements.txt
Django>=4.0,<5.0          
djangorestframework>=3.14 
psycopg2-binary>=2.9      
django-cors-headers>=3.13 
pandas
openpyxl 